fix(ux): address PR review -- date field error in SupplierInvoiceForm, extract useUnsavedChangesGuard composable
